don't treat velocity of 0 as a release
HaliKey v2 uses the velocity field as part of a special way of sending
timings to the computer, and in some cases that will result in it
sending a "note on" event with a velocity of 0. For a musical instrument
that would be silent, so the parsing code considered it the same as a
"note off". But since these aren't musical instruments we can bypass
that and simply look at the on/off directly.

Fixes #5.