Major keyer rewrite
This began as an attempt to fix an issue that turned out to be caused by
the lack of MIDI running status support (see the previous commit).
However, in increasing frustration I ended up rewriting the iambic keyer
from scratch. I think it's clearer now (it works more like the one in
nKEYER) and the Iambic A vs. B switch actually does something now.