handle "running status" in MIDI
MIDI uses bytes 0x80..0xff as "status bytes" (i.e. commands), and bytes
0x00..0x7f as value bytes (this is why most values are limited to
0-127). There's an optimization called "running status", in which, if
multiple events have the same status byte, the status byte is omitted,
and the device just sends a stream of data bytes (which we recognize
because they're < 0x80).

The MIDI library we're using doesn't do any of that for us so we have to
do it ourselves.