don't crash the MIDI parser on SysEx messages
1. System messages don't allow running status, so if we think the previous
message type was 0xF# and we think we see running status, something is wrong
and we should ignore data until we get back into sync.
2. SysEx messages don't have a fixed length, so the lookup table in the parser
returns a length of 0 for them (which results in the parser not advancing at
all). For SysEx we have to parse (or really discard, since we don't have any
use for them) until we see an end-of-SysEx marker, 0xF7.
3. Since SysEx messages are long they can be split across several callback
invocations, so we have to keep state of whether we were in the middle of
parsing one.