Debug logging to file as well
This makes the debug log more accessible to Windows users (stdout/stderr
aren't available outside of a debugger because NetKeyer is built as a
GUI app) and may come in handy to other users as well. The log file is
only created if the debug environment variable is set, so it won't cause
a problem for users who haven't specifically activated debugging.